restructure github-receiver code
