+++ apps/github-receiver/src/parsers/base.parser.ts
