+++ apps/github-receiver/src/parsers/pull-request-review-comment.parser.ts
@@ -1,5 +1,5 @@
 import { PullRequestReviewCommentEventData } from '@codeheroes/providers';
-import { PullRequestReviewCommentEvent } from '../interfaces/github.interfaces';
+import { PullRequestReviewCommentEvent } from '../core/interfaces/github.interfaces';
 import { GitHubParser } from './base.parser';
 
 export class PullRequestReviewCommentParser extends GitHubParser<
